feat: first part of refactor
