+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Interfaces/IBIWEditor.cs
@@ -1,4 +1,5 @@
 using DCL.Builder;
+using DCL.Controllers;
 
 public interface IBIWEditor
 {
@@ -7,4 +8,15 @@ public interface IBIWEditor
     void OnGUI();
     void Update();
     void LateUpdate();
+
+    /// <summary>
+    /// Open the editor with the scene that you pass. It must be ready to work correctly
+    /// </summary>
+    /// <param name="sceneToEdit"></param>
+    void EnterEditMode(IParcelScene sceneToEdit);
+
+    /// <summary>
+    /// Exits from the editor
+    /// </summary>
+    void ExitEditMode();
 }
