+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Interfaces/ISceneManager.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DCL.Builder
+{
+    public interface ISceneManager
+    {
+        void Initialize(IContext context);
+        void Dispose();
+        void Update();
+
+    }
+}
