+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Interfaces/ISceneManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 51f18fa968266054282dd71511fb95a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
