+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/SceneManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec7791768ac5a7b4cb6f9f2efa9697c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
