+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Tests/BIWMainControllerShould.cs
@@ -1,20 +1,11 @@
-using System;
 using System.Collections;
-using System.Collections.Generic;
-using System.IO;
 using DCL;
 using DCL.Builder;
 using DCL.Camera;
 using DCL.Controllers;
-using DCL.Helpers;
 using NSubstitute;
-using NSubstitute.Extensions;
-using NSubstitute.ReceivedExtensions;
 using NUnit.Framework;
-using Tests;
-using UnityEngine;
 using UnityGLTF;
-using Environment = DCL.Environment;
 
 public class BIWMainControllerShould : IntegrationTestSuite_Legacy
 {
@@ -26,49 +17,8 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         yield return base.SetUp();
         DataStore.i.builderInWorld.landsWithAccess.Set(new LandWithAccess[0]);
         mainController = new BuilderInWorldEditor();
-        BuilderInWorldEditor.BYPASS_LAND_OWNERSHIP_CHECK = true;
         apiSubstitute = Substitute.For<IBuilderAPIController>();
         mainController.Initialize(BIWTestUtils.CreateContextWithGenericMocks(apiSubstitute));
-        mainController.initialLoadingController.Dispose();
-        mainController.initialLoadingController = Substitute.For<IBuilderInWorldLoadingController>();
-        mainController.initialLoadingController.Configure().isActive.Returns(true);
-
-    }
-
-    [Test]
-    public void SetFlagProperlyWhenBuilderInWorldIsEntered()
-    {
-        // Arrange
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        // Act
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
-
-        // Assert
-        Assert.IsTrue(mainController.isBuilderInWorldActivated);
-    }
-
-    [Test]
-    public void SetFlagProperlyWhenBuilderInWorldIsExited()
-    {
-        // Arrange
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
-
-        // Act
-        mainController.ExitEditMode();
-
-        // Assert
-        Assert.IsFalse(mainController.isBuilderInWorldActivated);
     }
 
     [Test]
@@ -85,22 +35,6 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
     }
 
     [Test]
-    public void FindSceneToEdit()
-    {
-        // Arrange
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
-        CommonScriptableObjects.playerWorldPosition.Set(new Vector3(scene.sceneData.basePosition.x, 0, scene.sceneData.basePosition.y));
-
-        // Act
-        var sceneFound = mainController.FindSceneToEdit();
-
-        // Arrange
-        Assert.AreEqual(scene, sceneFound);
-    }
-
-    [Test]
     public void ControllerEnterEditMode()
     {
         // Arrange
@@ -108,13 +42,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         mainController.InitController(controller);
 
         // Act
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Assert
         controller.Received(1).EnterEditMode(scene);
@@ -126,14 +54,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         // Arrange
         IBIWController controller = Substitute.For<IBIWController>();
         mainController.InitController(controller);
-
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Act
         mainController.ExitEditMode();
@@ -148,14 +69,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         // Arrange
         IBIWController controller = Substitute.For<IBIWController>();
         mainController.InitController(controller);
-
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Act
         mainController.OnGUI();
@@ -170,14 +84,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         // Arrange
         IBIWController controller = Substitute.For<IBIWController>();
         mainController.InitController(controller);
-
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Act
         mainController.LateUpdate();
@@ -192,6 +99,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         // Arrange
         IBIWController controller = Substitute.For<IBIWController>();
         mainController.InitController(controller);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Act
         mainController.Update();
@@ -206,14 +114,7 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         // Arrange
         IBIWController controller = Substitute.For<IBIWController>();
         mainController.InitController(controller);
-
-        mainController.CatalogLoaded();
-        scene.CreateEntity("Test");
-
-        mainController.TryStartEnterEditMode(false, scene, "Test");
-        ParcelScene createdScene = (ParcelScene) Environment.i.world.sceneController.CreateTestScene(scene.sceneData);
-        createdScene.CreateEntity("TestEntity");
-        Environment.i.world.sceneController.SendSceneReady(scene.sceneData.id);
+        mainController.EnterEditMode(Substitute.For<IParcelScene>());
 
         // Act
         mainController.Update();
@@ -223,183 +124,13 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
     }
 
     [Test]
-    public void BIWExitWhenCharacterIsFarAway()
-    {
-        // Arrange
-        mainController.sceneToEdit = scene;
-        mainController.isBuilderInWorldActivated = true;
-        mainController.checkerInsideSceneOptimizationCounter = 60;
-        DCLCharacterController.i.characterPosition.unityPosition = Vector3.one * 9999;
-
-        // Act
-        mainController.Update();
-
-        // Assert
-        Assert.IsFalse(mainController.isBuilderInWorldActivated);
-    }
-
-    [Test]
-    public void ActivateLandAccessBackground()
-    {
-        // Arrange
-        var profile = UserProfile.GetOwnUserProfile();
-        profile.UpdateData(new UserProfileModel() { userId = "testId", ethAddress = "0x00" });
-
-        // Act
-        mainController.ActivateLandAccessBackgroundChecker();
-
-        // Assert
-        Assert.IsNotNull(mainController.updateLandsWithAcessCoroutine);
-    }
-
-    [Test]
-    public void RequestCatalog()
-    {
-        // Arrange
-        mainController.isCatalogRequested = false;
-        
-        ((Context)mainController.context).builderAPIController = Substitute.For<IBuilderAPIController>();
-        Promise<bool> resultOkPromise = new Promise<bool>();
-        mainController.context.builderAPIController.Configure().GetCompleteCatalog(Arg.Any<string>()).Returns(resultOkPromise);
-
-        // Act
-        mainController.GetCatalog();
-        resultOkPromise.Resolve(true);
-
-        // Assert
-        Assert.IsTrue(mainController.isCatalogRequested);
-    }
-
-    [Test]
-    public void ChangeEditModeByShortcut()
-    {
-        // Act
-        mainController.ChangeEditModeStatusByShortcut(DCLAction_Trigger.BuildEditModeChange);
-
-        // Assert
-        Assert.IsTrue(mainController.isWaitingForPermission);
-    }
-
-    [Test]
-    public void NewSceneAdded()
-    {
-        // Arrange
-        var mockedScene = Substitute.For<IParcelScene>();
-        mockedScene.Configure().sceneData.Returns(scene.sceneData);
-        mainController.sceneToEditId = scene.sceneData.id;
-
-        // Act
-        mainController.NewSceneAdded(mockedScene);
-
-        // Assert
-        Assert.AreSame(mainController.sceneToEdit, base.scene);
-    }
-
-    [Test]
-    public void UserHasPermission()
-    {
-        // Arrange
-        AddSceneToPermissions();
-
-        // Act
-        var result = mainController.UserHasPermissionOnParcelScene(scene);
-
-        // Assert
-        Assert.IsTrue(result);
-    }
-
-    [Test]
-    public void ReturnTrueWhenParcelSceneDeployedFromSDKIsCalled()
-    {
-        // Arrange
-        Parcel parcel = new Parcel();
-        parcel.x = base.scene.sceneData.basePosition.x;
-        parcel.y = base.scene.sceneData.basePosition.y;
-
-        Vector2Int parcelCoords = new Vector2Int(base.scene.sceneData.basePosition.x, base.scene.sceneData.basePosition.y);
-        Land land = new Land();
-        land.parcels = new List<Parcel>() { parcel };
-
-        LandWithAccess landWithAccess = new LandWithAccess(land);
-        Scene scene = new Scene();
-        scene.parcelsCoord = new Vector2Int[] { parcelCoords };
-        scene.deploymentSource = Scene.Source.SDK;
-
-        landWithAccess.scenes = new List<Scene>() { scene };
-        var lands = new LandWithAccess[]
-        {
-            landWithAccess
-        };
-        DataStore.i.builderInWorld.landsWithAccess.Set(lands);
-
-        // Act
-        var result = mainController.IsParcelSceneDeployedFromSDK(base.scene);
-
-        // Assert
-        Assert.IsTrue(result);
-    }
-
-    [Test]
-    public void CatalogReceived()
-    {
-        // Arrange
-        ((Context)mainController.context).builderAPIController = Substitute.For<IBuilderAPIController>();
-        Promise<bool> resultOkPromise = new Promise<bool>();
-        mainController.context.builderAPIController.Configure().GetCompleteCatalog(Arg.Any<string>()).Returns(resultOkPromise);
-        
-        // Act
-        mainController.GetCatalog();
-        resultOkPromise.Resolve(true);
-
-        // Assert
-        Assert.IsTrue(mainController.catalogAdded);
-    }
-
-    [Test]
-    public void CheckSceneToEditByShortcut()
-    {
-        // Arrange
-        mainController.sceneToEdit = scene;
-        AddSceneToPermissions();
-        ((Context)mainController.context).builderAPIController = Substitute.For<IBuilderAPIController>();
-        Promise<bool> resultOkPromise = new Promise<bool>();
-        mainController.context.builderAPIController.Configure().GetCompleteCatalog(Arg.Any<string>()).Returns(resultOkPromise);
-
-        // Act
-        mainController.CheckSceneToEditByShorcut();
-        resultOkPromise.Resolve(true);
-
-        // Assert
-        Assert.IsTrue(mainController.isEnteringEditMode);
-    }
-
-    [Test]
-    public void InitialLoadingControllerHideOnFloorLoaded()
-    {
-        // Arrange
-        mainController.initialLoadingController.Dispose();
-        mainController.initialLoadingController = Substitute.For<IBuilderInWorldLoadingController>();
-        mainController.initialLoadingController.Configure().isActive.Returns(true);
-
-        // Act
-        mainController.OnAllParcelsFloorLoaded();
-
-        // Assert
-        mainController.initialLoadingController.Received().Hide(Arg.Any<bool>(), Arg.Any<Action>());
-    }
-
-    [Test]
     public void StartExitModeScreenShot()
     {
         // Arrange
         mainController.Dispose();
         mainController = new BuilderInWorldEditor();
-        BuilderInWorldEditor.BYPASS_LAND_OWNERSHIP_CHECK = true;
         mainController.Initialize(BIWTestUtils.CreateContextWithGenericMocks(new BIWModeController(),
             new BIWSaveController(), SceneReferences.i));
-        mainController.initialLoadingController.Dispose();
-        mainController.initialLoadingController = Substitute.For<IBuilderInWorldLoadingController>();
-        mainController.initialLoadingController.Configure().isActive.Returns(true);
 
         BIWModeController modeController = (BIWModeController)mainController.modeController;
         BIWSaveController saveController = (BIWSaveController)mainController.saveController;
@@ -431,40 +162,6 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         mainController.context.editorContext.floorHandler.Received(1).CreateDefaultFloor();
     }
 
-    [Test]
-    public void ExitAfterTeleport()
-    {
-        // Arrange
-        mainController.sceneToEdit = scene;
-        mainController.isBuilderInWorldActivated = true;
-
-        // Act
-        mainController.ExitAfterCharacterTeleport(new DCLCharacterPosition());
-
-        // Assert
-        Assert.IsFalse(mainController.isBuilderInWorldActivated);
-    }
-
-    private void AddSceneToPermissions()
-    {
-        var parcel = new Parcel();
-        parcel.x = scene.sceneData.basePosition.x;
-        parcel.y = scene.sceneData.basePosition.y;
-
-        var land = new Land();
-        land.parcels = new List<Parcel>() { parcel };
-
-        var landWithAccess = new LandWithAccess(land);
-        landWithAccess.scenes = new List<Scene>();
-
-        var lands = new LandWithAccess[]
-        {
-            landWithAccess
-        };
-
-        DataStore.i.builderInWorld.landsWithAccess.Set(lands);
-    }
-
     protected override IEnumerator TearDown()
     {
         yield return new DCL.WaitUntil( () => GLTFComponent.downloadingCount == 0 );
@@ -473,7 +170,6 @@ public class BIWMainControllerShould : IntegrationTestSuite_Legacy
         DataStore.i.builderInWorld.landsWithAccess.Set(new LandWithAccess[0]);
         mainController.context.Dispose();
         mainController.Dispose();
-        BuilderInWorldEditor.BYPASS_LAND_OWNERSHIP_CHECK = false;
         yield return base.TearDown();
     }
 }
