+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Tests/BIWSceneManagerShould.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 898fc37f786415843b3e251703eb5d38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
